Cosmology now a full module with optional Phytorch
