+++ caustic/lenses/pseudo_jaffe.py
@@ -4,7 +4,7 @@ from typing import Any, Optional, Union
 import torch
 from torch import Tensor
 
-from ..cosmology import Cosmology
+from ..cosmology.base import Cosmology
 from ..utils import translate_rotate
 from .base import ThinLens
 from ..parametrized import unpack
