+++ caustic/lenses/singleplane.py
@@ -3,7 +3,7 @@ from typing import Any, Optional
 import torch
 from torch import Tensor
 
-from ..cosmology import Cosmology
+from ..cosmology.base import Cosmology
 from .base import ThinLens
 from ..parametrized import unpack
